Make content-first consistent components
