+++ content/index.md
@@ -1,16 +1,4 @@
 ---
-title: Under construction
+title: PostCSS - a tool for transforming CSS with JavaScript
 layout: HomePage
 ---
-
-PostCSS is a tool for transforming styles with JS plugins. These plugins can lint your CSS, support variables and mixins, transpile future CSS syntax, inline images, and more.
-
-PostCSS is used by industry leaders including Google, Twitter, Alibaba, and Shopify. The Autoprefixer PostCSS plugin is one of the most popular CSS processors.
-
-[Learn more about PostCSS](https://github.com/postcss/postcss)
-
-## Contribute to [PostCSS.org](https://github.com/postcss/postcss.org)?
-
-Please read through our [contributing guidelines](https://github.com/postcss/postcss.org/blob/master/CONTRIBUTING.md) to get started working on postcss.org and once you are done reading the guidelines, feel free to take a look at the [issue tracker](https://github.com/postcss/postcss.org/issues) from [Milestone 1.0](https://github.com/postcss/postcss.org/milestones/1.0) and start hacking away at something!
-
-There is a bunch of documentation on how to work on `postcss.org` in the `docs/` folder, [`Getting Started`](docs/getting-started.md) is a good place to begin!
