+++ web_modules/Button/index.css
@@ -1,4 +1,4 @@
-.button {
+.root {
  border: 1px solid CurrentColor;
  color: #333;
  cursor: pointer;
