--- web_modules/Hero/index.css
@@ -1,4 +1,0 @@
-.hero {
-  height: 350px;
-  background: #444;
-}
