+++ web_modules/layouts/HomePage/index.css
@@ -1,4 +1,9 @@
-.wrapper {
+.root {
   max-width: 50rem;
   margin: 0 auto;
+}
+
+.hero {
+  height: 350px;
+  background: #444;
 }
